requirements.txt
flask>=3.0
requests>=2.31
gunicorn>=21.2
openai>=1.0
python-dotenv>=1.0
